+++ src/main/java/com/team1816/lib/hardware/components/motor/LazyTalonFXDev.java
@@ -152,12 +152,12 @@ public class LazyTalonFXDev extends TalonFX implements IGreenMotorDev {
     }
 
     @Override
-    public void configOpenLoopRampRate(int secondsNeutralToFull) {
+    public void configOpenLoopRampRate(double secondsNeutralToFull) {
         super.configOpenloopRamp(secondsNeutralToFull);
     }
 
     @Override
-    public void configClosedLoopRampRate(int secondsNeutralToFull) {
+    public void configClosedLoopRampRate(double secondsNeutralToFull) {
         super.configClosedloopRamp(secondsNeutralToFull);
     }
