+++ src/main/java/com/team1816/lib/hardware/components/motor/LazyTalonSRXDev.java
@@ -129,12 +129,12 @@ public class LazyTalonSRXDev extends TalonSRX implements IGreenMotorDev{
     }
 
     @Override
-    public void configOpenLoopRampRate(int secondsNeutralToFull) {
+    public void configOpenLoopRampRate(double secondsNeutralToFull) {
         super.configOpenloopRamp(secondsNeutralToFull);
     }
 
     @Override
-    public void configClosedLoopRampRate(int secondsNeutralToFull) {
+    public void configClosedLoopRampRate(double secondsNeutralToFull) {
         super.configClosedloopRamp(secondsNeutralToFull);
     }
